Keep platform-dependent python extension filename (#11516)
By default, Python extension filenames include a platform-dependent
suffix. To facilitate the specification of the dependencies of the
Python extension file in CMake, this behavior was previously overwritten
by eliminating the suffix so that the filename was the same across all
platforms. Although CPython, the reference implementation of Python, can
load Python extensions with the new name without any problems, I
recently discovered that PyPy cannot do so properly. This PR reverses
the change, retains the original name, and updates the CMake and
Setuptools files to handle the different filenames.